+++ NEWS.md
@@ -1,4 +1,7 @@
 # RAthena (development version)
+## Feature:
+* Add support `dbplyr` 2.0.0 backend API.
+
 ## Bug Fix:
 * Ensure `dbGetQuery`, `dbExecute`, `dbSendQuery`, `dbSendStatement` work on older versions of `R` ([noctua # 170](https://github.com/DyfanJones/noctua/issues/170)). Thanks to @tyner for identifying issue.
 * Caching would fail when statement wasn't a character ([noctua # 171](https://github.com/DyfanJones/noctua/issues/171)). Thanks to @ramnathv for identifying issue.